增加评论逻辑后端

+++ src/main/java/life/royluo/community/community/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String publish(@PathVariable(name = "id")Integer id,
+    public String publish(@PathVariable(name = "id")Long id,
                           Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -62,7 +62,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model){
